Externalized previous Drug enum to DrugManager
